+++ crates/mesh-llm-commands/src/gpus/tune/output_types.rs
@@ -1,3 +1,7 @@
+use serde::Serialize;
+
+use super::*;
+
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub(crate) struct TuneTargetFailure {
     pub requested_input: String,
